+++ changelog.md
@@ -3,7 +3,7 @@
 ## NEXT
 
 - Added the method `from_dict` for all node and edges.
-- Changed behaviour of `cai_causal_graph.causal_graph.CausalGraph.add_node` when both `identifier` and (`time_lag`, `variable_name`) are provided. Now, if all are provided, the method will raise an error only if `identifier` is not
+- Changed behaviour of `cai_causal_graph.time_series.causal_graph.TimeSeriesCausalGraph.add_node` when both `identifier` and (`time_lag`, `variable_name`) are provided. Now, if all are provided, the method will raise an error only if `identifier` is not
   equal to `get_name_with_lag(time_lag, variable_name)`, that is, the correct name.
 
 ## 0.2.2